app/requirements.txt
flask>=3.0.0
flask-cors>=5.0.0
gunicorn>=22.0.0
minio>=7.2.0
pdfplumber>=0.11.0
python-docx>=1.1.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
requests>=2.31.0
transformers>=4.40.0
duckduckgo-search>=5.0.0
